Rename Dialogs::EncryptPassword step 1 - filenames
