+++ src/lib/y2partitioner/dialogs/encryption.rb
